inference/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.5
contourpy==1.0.7
cycler==0.11.0
dateparser==1.1.6
Flask==2.1.0
fonttools==4.38.0
frozenlist==1.3.3
google-api-core==2.10.2
google-auth==2.13.0
google-cloud-core==2.3.2
google-cloud-datastore==2.8.1
google-cloud-secret-manager==2.12.6
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio-status==1.50.0
grpcio==1.50.0
gym-notices==0.0.8
gym==0.24.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pip==21.2.4
proto-plus==1.22.1
protobuf==4.21.8
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
setuptools==58.1.0
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.14
websockets==10.4
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.9.0

-f https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
